+++ CHANGELOG.md
@@ -32,6 +32,10 @@ All notable changes to this project will be documented in this file.
 ## [2.3.0](https://github.com/uias/Tabman/releases/tag/2.3.0)
 Released on TBD
 
+#### Updated
+- Make `TMBarItem` `open`.
+- Make `title`, `image`, and `badgeValue` open on `TMBarItem`.
+
 ---
 ## [2.2.0](https://github.com/uias/Tabman/releases/tag/2.2.0)
 Released on 2019-01-14